+++ src/clean.rs
@@ -3,7 +3,7 @@ use anyhow::{Context, Result};
 use std::path::Path;
 
 pub fn clean_target_output_paths(target: &Target) -> Result<()> {
-    for output_path in &target.output_paths {
+    for output_path in &target.output.paths {
         clean_path(output_path)?;
     }