+++ src/domain.rs
@@ -9,8 +9,8 @@ pub struct Target {
     pub name: String,
     pub project: Project,
     pub dependencies: Vec<TargetId>,
-    pub input_paths: Vec<PathBuf>,
-    pub output_paths: Vec<PathBuf>,
+    pub input: Resources,
+    pub output: Resources,
     pub build: Option<String>,
     pub service: Option<String>,
 }
@@ -28,4 +28,23 @@ impl fmt::Display for Target {
 pub struct Project {
     pub dir: PathBuf,
     pub name: Option<String>,
+}
+
+#[derive(Clone, Debug)]
+pub struct Resources {
+    pub paths: Vec<PathBuf>,
+    pub cmds: Vec<String>,
+}
+
+impl Resources {
+    pub fn new() -> Self {
+        Self {
+            paths: vec![],
+            cmds: vec![],
+        }
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.paths.is_empty() && self.cmds.is_empty()
+    }
 }