+++ src/engine/builder.rs
@@ -28,7 +28,7 @@ pub fn build_target(target: &Target, termination_events: Receiver<()>) -> Result
                 recv(ticks) -> _ => {
                     if let Some(exit_status) = build_process.try_wait()? {
                         if !exit_status.success() {
-                            return Err(anyhow::anyhow!("Build failed for target {} ({})", target.name, exit_status));
+                            return Err(anyhow::anyhow!("Build failed for target {} ({})", target, exit_status));
                         }
                         let target_build_duration = target_start.elapsed();
                         log::info!(
@@ -40,8 +40,8 @@ pub fn build_target(target: &Target, termination_events: Receiver<()>) -> Result
                     }
                 },
                 recv(termination_events) -> _ => {
-                    process::kill_and_wait(&mut build_process).with_context(|| format!("Failed to kill build process for {}", target.name))?;
-                    return Err(anyhow::anyhow!("Build cancelled for target {}", target.name));
+                    process::kill_and_wait(&mut build_process).with_context(|| format!("Failed to kill build process for {}", target))?;
+                    return Err(anyhow::anyhow!("Build cancelled for target {}", target));
                 },
             }
         }