+++ src/main.rs
@@ -3,15 +3,17 @@ mod cli;
 mod config;
 mod domain;
 mod engine;
+mod work_dir;
 
 use anyhow::{Context, Result};
 use clean::clean_target_output_paths;
 use config::{ir, yaml};
 use crossbeam::channel::{unbounded, Sender};
-use engine::incremental::{remove_checksum_dir, remove_target_checksums};
+use engine::incremental::storage::delete_saved_env_state;
 use engine::Engine;
 use std::convert::TryInto;
 use std::path::PathBuf;
+use work_dir::remove_work_dir;
 
 #[cfg(all(not(target_env = "msvc"), target_pointer_width = "64"))]
 use jemallocator::Jemalloc;
@@ -53,9 +55,9 @@ fn main() -> Result<()> {
 
     if arg_matches.is_present(cli::arg::CLEAN) {
         if has_requested_targets {
-            targets.iter().try_for_each(remove_target_checksums)?;
+            targets.iter().try_for_each(delete_saved_env_state)?;
         } else {
-            project_dirs.into_iter().try_for_each(remove_checksum_dir)?;
+            project_dirs.into_iter().try_for_each(remove_work_dir)?;
         }
 
         targets.iter().try_for_each(clean_target_output_paths)?;