+++ tests/integ.rs
@@ -1,4 +1,3 @@
-use assert_cmd::assert::Assert;
 use assert_cmd::prelude::*;
 use predicates::prelude::*;
 use std::ffi;
@@ -6,28 +5,32 @@ use std::process::Command;
 
 #[test]
 fn circular_dependency() {
-    assert_zinoma("circular_dependency", &["target_1"])
+    zinoma_command("circular_dependency", &["target_1"])
+        .assert()
         .failure()
         .stderr(predicate::str::contains("Circular dependency"));
 }
 
 #[test]
 fn imports() {
-    assert_zinoma("imports", &["target_2"])
+    zinoma_command("imports", &["target_2"])
+        .assert()
         .success()
         .stdout(predicate::str::contains("This is target 1"));
 }
 
 #[test]
 fn named_project() {
-    assert_zinoma("named_project", &["target_3", "my_project::target_3"])
+    zinoma_command("named_project", &["target_3", "my_project::target_3"])
+        .assert()
         .success()
         .stdout(predicate::str::contains("This is target 1"));
 }
 
 #[test]
 fn non_matching_imported_project_name() {
-    assert_zinoma("non_matching_imported_project_name", &["--clean"])
+    zinoma_command("non_matching_imported_project_name", &["--clean"])
+        .assert()
         .failure()
         .stderr(predicate::str::contains(
             "Failed to import incorrect_subproject_name",
@@ -39,7 +42,8 @@ fn non_matching_imported_project_name() {
 
 #[test]
 fn import_project_with_no_name() {
-    assert_zinoma("import_project_with_no_name", &["--clean"])
+    zinoma_command("import_project_with_no_name", &["--clean"])
+        .assert()
         .failure()
         .stderr(predicate::str::contains("Failed to import noname"))
         .stderr(predicate::str::contains(
@@ -49,24 +53,50 @@ fn import_project_with_no_name() {
 
 #[test]
 fn invalid_project_name() {
-    assert_zinoma("invalid_project_name", &["--clean"])
+    zinoma_command("invalid_project_name", &["--clean"])
+        .assert()
         .failure()
         .stderr(predicate::str::contains(":::: is not a valid project name"));
 }
 
 #[test]
 fn root_input_path() {
-    assert_zinoma("root_input_path", &["--clean", "print_source"])
+    zinoma_command("root_input_path", &["--clean", "print_source"])
+        .assert()
         .success()
         .stdout(predicate::str::contains("Content of my source file"));
 
-    assert_zinoma("root_input_path", &["print_source"])
+    zinoma_command("root_input_path", &["print_source"])
+        .assert()
         .success()
         .stdout(predicate::str::contains("Content of my source file").not())
         .stderr(predicate::str::contains("Build skipped (Not Modified)"));
 }
 
-fn assert_zinoma<I, S>(integ_test_dir_name: &str, args: I) -> Assert
+#[test]
+fn cmd_stdout_input() {
+    let integ_test_dir_name = if cfg!(windows) {
+        "cmd_stdout_input_windows"
+    } else if cfg!(target_os = "macos") {
+        "cmd_stdout_input_macos"
+    } else {
+        "cmd_stdout_input"
+    };
+
+    zinoma_command(integ_test_dir_name, &["--clean", "changing", "stable"])
+        .assert()
+        .success()
+        .stderr(predicate::str::contains("changing - Build success"))
+        .stderr(predicate::str::contains("stable - Build success"));
+
+    zinoma_command(integ_test_dir_name, &["changing", "stable"])
+        .assert()
+        .success()
+        .stderr(predicate::str::contains("changing - Build success"))
+        .stderr(predicate::str::contains("stable - Build skipped"));
+}
+
+fn zinoma_command<I, S>(integ_test_dir_name: &str, args: I) -> Command
 where
     I: IntoIterator<Item = S>,
     S: AsRef<ffi::OsStr>,
@@ -75,5 +105,5 @@ where
     cmd.arg("-p")
         .arg(format!("tests/integ/{}", integ_test_dir_name))
         .args(args);
-    cmd.assert()
+    cmd
 }